Move managerReducer tests to mimic source code directory structure
